% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzip_conda.R
\name{bgzip_conda}
\alias{bgzip_conda}
\title{Run bgzip: conda}
\usage{
bgzip_conda(
  fullSS_path,
  bgz_file = construct_tabix_path(path = fullSS_path),
  chrom_col,
  start_col,
  end_col,
  comment_char = NULL,
  conda_env = "echoR",
  verbose = TRUE
)
}
\description{
Support function for \link[echotabix]{run_bgzip}.
}
\keyword{internal}
