% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzip_rsamtools.R
\name{bgzip_rsamtools}
\alias{bgzip_rsamtools}
\title{Run bgzip: Rsamtools}
\usage{
bgzip_rsamtools(fullSS_path, bgz_file, force_new = TRUE, verbose = TRUE)
}
\description{
Support function for \link[echotabix]{run_bgzip}.
}
\keyword{internal}
