% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_outputs.R
\name{construct_outputs}
\alias{construct_outputs}
\title{Construct outputs}
\usage{
construct_outputs(
  outputs,
  command = NULL,
  path = NULL,
  data = NULL,
  verbose = TRUE
)
}
\arguments{
\item{outputs}{\describe{
\item{"command"}{Text string of the command (without executing it.)}
\item{"path"}{Path to the saved data file.}
\item{"data"}{The resulting data in \link[data.table]{data.table} format.}
}}

\item{command}{Whether to return the command.}

\item{path}{Whether to return the path.}

\item{data}{Whether to return the data.}

\item{verbose}{Print messages.}
}
\description{
Construct a list of requested outputs.
}
\keyword{internal}
