% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_subset_vcf_name.R
\name{construct_subset_vcf_name}
\alias{construct_subset_vcf_name}
\title{Construct the path to vcf subset}
\usage{
construct_subset_vcf_name(dat, vcf_name = NULL, locus_dir, whole_vcf = FALSE)
}
\description{
\code{
locus_dir <- echodata::locus_dir
BST1 <- echodata::BST1
vcf_subset <- construct_subset_vcf_name(
    dat = BST1,
    locus_dir = locus_dir,
    vcf_name = "1KGlocal"
)
}
}
\seealso{
Other LD: 
\code{\link{get_locus_vcf_folder}()},
\code{\link{query_vcf}()}
}
\concept{LD}
\keyword{internal}
