% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tabix_path.R
\name{construct_tabix_path}
\alias{construct_tabix_path}
\title{Construct tabix path}
\usage{
construct_tabix_path(path, study_dir = NULL)
}
\arguments{
\item{path}{Path to file.}

\item{study_dir}{[optional] Path to study-specific subfolder.}
}
\description{
Given some summary stats file, 
construct a name for the resulting bgzip-compressed tabix file.
}
\examples{
bgz_file = echotabix::construct_tabix_path(path = "mysumstatsfile.tsv.gz") 
}
