% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{Fill NAs in an LD matrix}
\usage{
fill_na(LD_matrix, fillNA = 0, verbose = FALSE)
}
\arguments{
\item{LD_matrix}{LD matrix}

\item{fillNA}{Fill all NAs in the \code{LD_matrix} with \code{fillNA} value.}

\item{verbose}{Print messages.}
}
\description{
Trickier than it looks.
}
\details{
\code{
BST1_LD_matrix <- echodata::BST1_LD_matrix
LD_matrix <- fill_na(BST1_LD_matrix)
}
}
\keyword{internal}
