% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_vcf_samples.R
\name{filter_vcf_query_samples}
\alias{filter_vcf_query_samples}
\title{Filter VCF query samples}
\usage{
filter_vcf_query_samples(gr, samples = character(), verbose = TRUE)
}
\arguments{
\item{gr}{A \link[GenomicRanges]{GRanges} object 
generated by \link[echotabix]{construct_query}.}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{verbose}{Print messages.}
}
\value{
\link[GenomicRanges]{GRanges} or
\link[VariantAnnotation]{ScanVcfParam} object.
}
\description{
Only include certain samples in a VCF query. 
If at least one sample is supplied, returns
 a \link[VariantAnnotation]{ScanVcfParam}.
Otherwise, returns the original \link[GenomicRanges]{GRanges} object
 (\code{gr}).
}
\keyword{internal}
