% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chain_file.r
\name{get_chain_file}
\alias{get_chain_file}
\title{Download chain file for liftover}
\source{
\href{https://hgdownload.cse.ucsc.edu/goldenpath/hg19/liftOver/}{
UCSC chain files}
}
\usage{
get_chain_file(
  build_conversion = c("hg38ToHg19", "hg19ToHg38"),
  save_dir = tools::R_user_dir(package = "echotabix", which = "cache"),
  timeout = 60 * 5,
  verbose = TRUE
)
}
\arguments{
\item{build_conversion}{Converting from what build to what? 
"hg38ToHg19" or "hg19ToHg38"}

\item{save_dir}{Where is the chain file saved? Default is a temp directory}

\item{verbose}{extra messages printed? Default is TRUE}
}
\value{
Loaded chain file for liftover.
}
\description{
Download chain file for liftover
}
\seealso{
Other liftover functions: 
\code{\link{liftover}()}
}
\concept{liftover functions}
\keyword{internal}
