% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locus_vcf_folder.R
\name{get_locus_vcf_folder}
\alias{get_locus_vcf_folder}
\title{Get VCF storage folder}
\usage{
get_locus_vcf_folder(locus_dir = NULL)
}
\description{
Get VCF storage folder
}
\seealso{
Other LD: 
\code{\link{construct_subset_vcf_name}()}
}
\concept{LD}
\keyword{internal}
