% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_to_string.R
\name{granges_to_string}
\alias{granges_to_string}
\title{GRanges to string}
\usage{
granges_to_string(
  gr,
  pos_sep = "-",
  chrom_sep = ":",
  ranges_sep = ",",
  verbose = TRUE
)
}
\arguments{
\item{gr}{\link[GenomicRanges]{GRanges} object.}

\item{pos_sep}{Character to separate start/end genomic positions with.}

\item{chrom_sep}{Character to separate chromosome name 
from start/end positions.}

\item{ranges_sep}{Character to separate each genomic range with.}

\item{verbose}{Print messages.}
}
\value{
A concatenated string of coordinates.
}
\description{
Convert a \link[GenomicRanges]{GRanges} object 
to a concatenated string of coordinates 
(e.g. "chr4:70000-90000,chr10:200-150001"). 
This can be used for specifying
which regions you want to query (e.g. when using \code{tabix}).
}
\examples{
\dontrun{
gr1 <- echotabix::construct_query(query_dat = echodata::BST1)
gr2 <- echotabix::construct_query(query_dat = echodata::LRRK2)
gr <- suppressWarnings(c(gr1, gr2))

string <- echotabix::granges_to_string(gr=gr)
}
}
