% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_conda.R
\name{index_conda}
\alias{index_conda}
\title{Tabix-index a file: conda}
\usage{
index_conda(
  bgz_file,
  chrom_i,
  start_i,
  end_i,
  comment_char,
  skip_lines = 0,
  force_new = FALSE,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\arguments{
\item{bgz_file}{Path to a file that has been compressed with \code{bgzip}
(e.g. via \link[echotabix]{run_bgzip}).}

\item{comment_char}{Comment character denoting which row contains the column 
names (e.g. "#CHR" or "SNP").}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{verbose}{Print messages.}
}
\description{
Tabix-index a tabular summary statistics file.
}
