% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_variantannotation.R
\name{index_variantannotation}
\alias{index_variantannotation}
\title{Tabix-index file: VariantAnnotation}
\source{
Borrowed function from 
\href{https://github.com/RajLabMSSM/echotabix/blob/main/R/convert.R}{
echotabix}.
}
\usage{
index_variantannotation(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to VCF.}

\item{verbose}{Print messages.}
}
\value{
Path to tabix-indexed tabular file
}
\description{
Convert summary stats file to tabix format
}
\seealso{
Other tabix: 
\code{\link{convert_and_query}()}
}
\concept{tabix}
\keyword{internal}
