% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_chrom_type.R
\name{infer_chrom_type}
\alias{infer_chrom_type}
\title{Infer chromosome type}
\usage{
infer_chrom_type(chrom = NULL, path = NULL, chrom_col = "CHR", verbose = TRUE)
}
\arguments{
\item{chrom}{A value (character, numeric, or integer) 
giving an example of the chromosome column.}

\item{path}{Path to stored file.}

\item{chrom_col}{The name of the chromosome column in the file.}

\item{verbose}{Print messages.}
}
\description{
In a given vector or stored table, 
determine whether chromosome type is of the format \code{"chr1" }
(returns \code{TRUE}) or \code{1} (returns \code{FALSE}).
}
\examples{
echotabix::infer_chrom_type(chrom="chr1")
echotabix::infer_chrom_type(chrom=1)
}
