% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.r
\name{liftover}
\alias{liftover}
\title{Genome build liftover}
\source{
\href{https://doi.org/doi:10.18129/B9.bioc.liftOver}{liftOver}

\href{https://hgdownload.cse.ucsc.edu/goldenpath/hg19/liftOver/}{
UCSC chain files}
}
\usage{
liftover(
  dat,
  query_genome,
  target_genome,
  query_chrom_col = "CHR",
  query_start_col = "POS",
  query_end_col = query_start_col,
  pos_prefix = "POS",
  as_granges = FALSE,
  style = "NCBI",
  verbose = TRUE
)
}
\arguments{
\item{dat}{\link[data.table]{data.table} of GWAS/QTL summary statistics.}

\item{query_genome}{name of the reference genome used for the GWAS
(e.g. "GRCh37" or "GRCh38"). Argument is case-insensitive.}

\item{target_genome}{name of the reference genome to convert to
("GRCh37" or "GRCh38"). This will only occur if the \code{query_genome} does
not match the \code{target_genome}.}

\item{query_chrom_col}{Name of the chromosome column in 
\code{query_dat} (e.g. "CHR").}

\item{query_start_col}{Name of the starting genomic position
column in \code{query_dat} (e.g. "POS","start").}

\item{query_end_col}{Name of the ending genomic position
column in \code{query_dat} (e.g. "POS","end"). 
Can be the same as \code{query_start_col} when \code{query_dat}
 only contains SNPs that span 1 base pair (bp) each.}

\item{pos_prefix}{After the \code{start_col}\code{end_col}
has been lifted over, how should these columns be named?}

\item{as_granges}{Return results as \link[GenomicRanges]{GRanges} 
instead of a \link[data.table]{data.table} (default: \code{FALSE}).}

\item{style}{Style to return \link[GenomicRanges]{GRanges} object in
(e.g.  "NCBI" = 4; "UCSC" = "chr4";) (default: \code{"NCBI"}).}

\item{verbose}{Print messages.}
}
\value{
Lifted summary stats in \code{data.table} 
or \link[GenomicRanges]{GRanges} format.
}
\description{
Transfer genomic coordinates from one genome build to another.
}
\examples{
\dontrun{
dat <- echodata::BST1
#### hg19 ==> hg38 ####
dat_lifted <- liftover(
    dat = dat,
    query_genome = "hg19",
    target_genome = "hg38"
)
}
}
\seealso{
Other liftover functions: 
\code{\link{get_chain_file}()}
}
\concept{liftover functions}
