% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table.R
\name{query_table}
\alias{query_table}
\title{Query a tabix file}
\usage{
query_table(
  target_path,
  target_index = paste0(target_path, ".tbi"),
  query_granges,
  query_method = c("rsamtools", "seqminer", "conda"),
  local = NULL,
  overlapping_only = FALSE,
  query_save = TRUE,
  save_path = tempfile(fileext = "tsv.gz"),
  cleanup_tbi = TRUE,
  conda_env = "echoR_mini",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{target_index}{Tabix index file for \code{target_path}.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{query_method}{Method used for querying. 
See \link[echotabix]{query} for available options.}

\item{local}{Whether \code{target_path} is stored locally or 
on a remote server/website.
By default (\code{NULL}) will infer local status and 
use the appropriate \code{query_method}.}

\item{overlapping_only}{Remove variants that do not overlap with the 
positions in \code{query_dat}.}

\item{query_save}{Whether to save the queried data subset.}

\item{save_path}{File path to save query subset to (as table).}

\item{cleanup_tbi}{Remove local copies of tabix index file (\emph{.tbi}) 
after completing queries.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{nThread}{Number of threads to use.}

\item{verbose}{Print messages.}
}
\value{
\code{data.table} with the queried subset of genomic data.
}
\description{
Query by genomic coordinates.
}
\examples{
\dontrun{
query_dat <- echodata::BST1

#### local ####
target_path <- echodata::example_fullSS()
tabix_files <- echotabix::convert(target_path = target_path,
                                  start_col = "BP")
query_res <- echotabix::query_table(
    target_path = tabix_files$path,
    query_dat = query_dat)

#### remote ####
target_path <- file.path(
    "https://egg2.wustl.edu/roadmap/data/byFileType",
    "chromhmmSegmentations/ChmmModels/coreMarks/jointModel/final",
    "E099_15_coreMarks_dense.bed.bgz"
)
query_res2 <- echotabix::query_table(
    target_path = target_path,
    query_granges = query_dat)
}
}
\seealso{
Other tabix functions: 
\code{\link{construct_tabix_path}()},
\code{\link{construct_vcf_path}()},
\code{\link{convert}()},
\code{\link{index}},
\code{\link{query_vcf}()},
\code{\link{read_bgz}()},
\code{\link{run_bgzip}()}
}
\concept{tabix functions}
