% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table_conda.R
\name{query_table_conda}
\alias{query_table_conda}
\title{Query tabular: conda}
\usage{
query_table_conda(
  target_path,
  query_dat,
  query_granges = construct_query(query_dat = query_dat, query_chrom_col = "CHR",
    query_start_col = "POS", query_snp_col = "SNP"),
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{query_dat}{Variant-level summary statistics.}

\item{verbose}{Print messages.}
}
\description{
Uses a conda-based installation of tabix instead of compiled 
\code{C} code from \pkg{Rhtslib}.
}
\keyword{internal}
