% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table_genomicfiles.R
\name{query_table_genomicfiles}
\alias{query_table_genomicfiles}
\title{Query tabular: \pkg{GenomicFiles}}
\usage{
query_table_genomicfiles(
  target_path,
  target_index = paste0(target_path, ".tbi"),
  query_granges,
  yieldSize = NA_character_,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{target_index}{Tabix index file for \code{target_path}.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{verbose}{Print messages.}
}
\description{
\pkg{GenomicFiles} can query both local and remote files.
}
\keyword{internal}
