% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table_rsamtools.R
\name{query_table_rsamtools}
\alias{query_table_rsamtools}
\title{Query tabular: \pkg{Rsamtools}}
\source{
\href{https://github.com/Bioconductor/Rsamtools/issues/8}{
 Bugs in \code{Rsamtools::scanTabix}}
}
\usage{
query_table_rsamtools(
  target_path,
  query_dat,
  query_granges = construct_query(query_dat = query_dat, query_chrom_col = "CHR",
    query_start_col = "POS", query_snp_col = "SNP"),
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{query_dat}{Variant-level summary statistics.}

\item{verbose}{Print messages.}
}
\description{
\pkg{Rsamtools} can query both local and remote files. 
 However, it seems unable to read the header of 
 tabix files created with \pkg{seqminer}.
}
\keyword{internal}
