% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tabular.R
\name{query_tabular}
\alias{query_tabular}
\title{Query a tabix file}
\usage{
query_tabular(fullSS_tabix, chrom, start_pos, end_pos, verbose = TRUE)
}
\description{
Query by genomic coordinates.
}
\examples{
\dontrun{
data("BST1")
fullSS_path <- echolocatoR::example_fullSS()
fullSS_tabix <- convert(fullSS_path = fullSS_path, position_col = "BP")
tab <- query_tabular(
    fullSS_tabix = fullSS_tabix,
    chrom = BST1$CHR[1],
    start_pos = min(BST1$POS),
    end_pos = max(BST1$POS)
)
}
}
\concept{query functions}
