% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tabular_seqminer.R
\name{query_tabular_seqminer}
\alias{query_tabular_seqminer}
\title{query_tabular_seqminer}
\usage{
query_tabular_seqminer(fullSS_tabix, chrom, start_pos, end_pos, verbose = TRUE)
}
\description{
\pkg{seqminer} is faster than \code{Rsamtools} but doesn't work
for remote files. Also, \pkg{seqminer} causes errors on Windows
and doesn't seem to be actively maintained anymore,
so should be avoided where possible.
}
\keyword{internal}
