% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf_conda.R
\name{query_vcf_conda}
\alias{query_vcf_conda}
\title{Query VCF: \pkg{conda}}
\usage{
query_vcf_conda(
  target_path,
  query_granges,
  samples = character(),
  download_index = FALSE,
  query_save = FALSE,
  save_path = NULL,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to local VCF file or remote URL.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{download_index}{Whether to download the index when querying.
Corresponds to the \code{-D} argument in \pkg{tabix}.}

\item{query_save}{Whether to save the results of the query on disk. 
\emph{Note}: Writing to disk can take some time.}

\item{save_path}{File path to save query subset to (as VCF).}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{verbose}{Print messages.}
}
\value{
\link[VariantAnnotation]{CollapsedVCF} object.
}
\description{
Query a subset of a VCF file (remote or local) 
using \code{tabix} via \pkg{echoconda}.
\strong{Advantages:}
\itemize{
\item Fast.
\item Not dependent on any R packages.
}
}
\keyword{internal}
