% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf_rtracklayer.R
\name{query_vcf_rtracklayer}
\alias{query_vcf_rtracklayer}
\title{Query VCF: \pkg{rtracklayer}}
\source{
\href{https://github.com/lawremi/rtracklayer/issues/62}{
Inconsistencies with  \pkg{rtracklayer}}

\code{
BST1 <- echodata::BST1
query_dat <- BST1[seq(1, 50), ]
target_path <- paste(
    "ftp://ftp-trace.ncbi.nih.gov/1000genomes/ftp/release/20110521/",
    "ALL.chr4.phase1_release_v3.20101123.snps_indels_svs.genotypes.vcf.gz",
    sep="/"
)
vcf <- echotabix:::query_vcf_rtracklayer(
    target_path = target_path,
    query_granges = query_dat)
}
}
\usage{
query_vcf_rtracklayer(
  target_path,
  target_index = paste0(target_path, ".tbi"),
  query_granges,
  samples = character(),
  query_save = FALSE,
  save_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to local VCF file or remote URL.}

\item{target_index}{Tabix index file for \code{target_path}.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{query_save}{Whether to save the results of the query on disk. 
\emph{Note}: Writing to disk can take some time.}

\item{save_path}{File path to save query subset to (as VCF).}

\item{verbose}{Print messages.}
}
\value{
Raw text?
}
\description{
Query a subset of a VCF file (remote or local) 
using \code{rtracklayer::import}, which is essentially just a wrapper for 
\link[Rsamtools]{scanTabix}.
\strong{Advantages of \pkg{rtracklayer}:}
\itemize{
\item None to speak of.
}
\strong{Disadvantages of \pkg{rtracklayer}:}
\itemize{
\item Unable to query a subset of samples,
unlike \link[VariantAnnotation]{scanVcf}.
\item Unable to return results as a structured
\link[VariantAnnotation]{CollapsedVCF} object.
}
}
\keyword{internal}
