% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bgz.R
\name{read_bgz}
\alias{read_bgz}
\title{Read bgz}
\usage{
read_bgz(
  path,
  method = c("data.table", "utils"),
  nrows = NULL,
  header = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{path}{Path to bgz-compressed file.}

\item{method}{Method to use to import bgz file.}

\item{nrows}{ The maximum number of rows to read. Unlike \code{read.table}, you do not need to set this to an estimate of the number of rows in the file for better speed because that is already automatically determined by \code{fread} almost instantly using the large sample of lines. \code{nrows=0} returns the column names and typed empty columns determined by the large sample; useful for a dry run of a large file or to quickly check format consistency of a set of files before starting to read any of them. }

\item{header}{ Does the first data line contain column names? Defaults according to whether every non-empty field on the first data line is type character. If so, or TRUE is supplied, any empty column names are given a default name. }

\item{verbose}{ Be chatty and report timings? }

\item{...}{Additional arguments passed to 
\link[data.table]{fread} or \link[utils]{read.delim2}.}
}
\description{
Read a compressed bgz file into R.
}
\details{
NOTE: Packages like \pkg{Rsamtools} and \pkg{seqminer} aren't great for 
this because they require that the bgz file it also indexed already.
}
\examples{
\dontrun{
tmp <- tempfile(fileext = ".tsv.gz")
data.table::fwrite(echodata::BST1, file = tmp, sep = "\t")
path <- Rsamtools::bgzip(file = tmp, overwrite=TRUE)
dat <- echotabix::read_bgz(path=path, method="utils")
}
}
\seealso{
Other tabix functions: 
\code{\link{construct_tabix_path}()},
\code{\link{construct_vcf_path}()},
\code{\link{convert}()},
\code{\link{index}},
\code{\link{query_table}()},
\code{\link{query_vcf}()},
\code{\link{run_bgzip}()}
}
\concept{tabix functions}
