% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtslib_warning.R
\name{rhtslib_warning}
\alias{rhtslib_warning}
\title{\pkg{Rsamtools} warning}
\source{
\href{https://github.com/Bioconductor/Rsamtools/issues/33#}{
Rsamtools/Rhtslib updates}

\href{https://github.com/Bioconductor/Rhtslib/issues/4}{
Rhtslib<1.99.2 (which Rsamtools and seqminer depend on for tabix) 
is very out of date (uses htslib 1.7 vs. 1.15)}.

\href{https://github.com/Bioconductor/Rsamtools}{
Rsamtools: GitHub}

Rsamtools: Bioconductor
(\doi{10.18129/B9.bioc.Rsamtools})
}
\usage{
rhtslib_warning(
  rhtslib_pkgs = c("variantannotation", "rsamtools", "rtracklayer"),
  method = NULL,
  verbose = TRUE
)
}
\arguments{
\item{rhtslib_pkgs}{List of R packages that depend on \code{Rhtslib}.}

\item{method}{Method requested.}

\item{verbose}{Print messages.}
}
\value{
Whether the Bioc version is invalid for this function.
}
\description{
Warn users about outdated version of \code{htslib} 
used by older versions of the \pkg{Rhtslib} R package.
}
\keyword{internal}
