% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_tbi.R
\name{rm_tbi}
\alias{rm_tbi}
\title{Remove tbi}
\usage{
rm_tbi(path = ".", pattern = "\\\\.tbi$")
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link[base]{getwd}()}.  Tilde
    expansion (see \code{\link[base]{path.expand}}) is performed.  Missing
    values will be ignored.  Elements with a marked encoding will
    be converted to the native encoding (and if that fails, considered
    non-existent).}

\item{pattern}{an optional \link[base]{regular expression}.  Only file names
    which match the regular expression will be returned.}
}
\value{
None
}
\description{
Remove local copies of tabix index file (\emph{.tbi}) 
after completing queries.
}
\keyword{internal}
