% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsamtools_warning.R
\name{rsamtools_warning}
\alias{rsamtools_warning}
\title{\pkg{Rsamtools} warning}
\source{
\href{https://github.com/Bioconductor/Rhtslib/issues/4}{
Rhtslib (which Rsamtools and seqminer depend on for tabix) 
is very out of date (1.7 vs. 1.15)}.

\href{https://github.com/Bioconductor/Rsamtools}{
Rsamtools: GitHub}

\href{https://doi.org/doi:10.18129/B9.bioc.Rsamtools}{
Rsamtools: Bioconductor}
}
\usage{
rsamtools_warning(old_version = "1.7", latest_version = "1.15", verbose = TRUE)
}
\description{
Warn users about outdated version of \code{htslib} used by \pkg{Rsamtools}.
}
\keyword{internal}
