% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bgzip.R
\name{run_bgzip}
\alias{run_bgzip}
\title{Run bgzip}
\usage{
run_bgzip(
  target_path,
  chrom_col,
  start_col,
  end_col = start_col,
  comment_char = NULL,
  bgz_file = construct_tabix_path(target_path = target_path),
  sort_rows = TRUE,
  force_new = TRUE,
  method = c("rsamtools", "conda"),
  conda_env = "echoR_mini",
  validate = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to full GWAS/QTL summary statistics file.}

\item{chrom_col}{Name of the chromosome column 
in the \code{target_path} file.}

\item{start_col}{Name of the genomic start position column 
in the \code{target_path} file.}

\item{end_col}{Name of the genomic end position column 
in the \code{target_path} file.}

\item{comment_char}{A single character which, when present as 
the first character in a line, indicates that the line is to be omitted 
from indexing.}

\item{bgz_file}{Path to resulting bgz-compressed file after
\code{target_path} has been converted to tabix format.}

\item{sort_rows}{Sort rows by genomic coordinates.}

\item{force_new}{Force the creation of a 
new bgzip file (\emph{.bgz}) and a new tabix index file (\emph{.tbi}).}

\item{method}{Method used to bgzip-compress the \code{target_path} file.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{validate}{Check that the bgzip file exists and can be read in as a 
\link[data.table]{data.table}.}

\item{verbose}{Print messages.}
}
\description{
Compress a file using bgzip.
}
\examples{
\dontrun{
#### Example with full data ####
# tmp <- echodata::example_fullSS()
#### Example with single locus ####
dat <- echodata::BST1
tmp <- tempfile()
data.table::fwrite(dat, tmp)

bgz_file <-  echotabix::run_bgzip(target_path=tmp,
                                 chrom_col="CHR",
                                 start_col="BP")
}
}
\seealso{
Other tabix functions: 
\code{\link{construct_tabix_path}()},
\code{\link{construct_vcf_path}()},
\code{\link{convert}()},
\code{\link{index}},
\code{\link{query_table}()},
\code{\link{query_vcf}()},
\code{\link{read_bgz}()}
}
\concept{tabix functions}
