% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bgzip_conda.R
\name{run_bgzip_conda}
\alias{run_bgzip_conda}
\title{Run bgzip: conda}
\usage{
run_bgzip_conda(
  target_path,
  bgz_file = construct_tabix_path(target_path = target_path),
  chrom_col,
  start_col,
  end_col,
  comment_char = NULL,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to full GWAS/QTL summary statistics file.}

\item{bgz_file}{Path to resulting bgz-compressed file after
\code{target_path} has been converted to tabix format.}

\item{chrom_col}{Name of the chromosome column 
in the \code{target_path} file.}

\item{start_col}{Name of the genomic start position column 
in the \code{target_path} file.}

\item{end_col}{Name of the genomic end position column 
in the \code{target_path} file.}

\item{comment_char}{A single character which, when present as 
the first character in a line, indicates that the line is to be omitted 
from indexing.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{verbose}{Print messages.}
}
\description{
Support function for \link[echotabix]{run_bgzip}.
}
\keyword{internal}
