% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bgzip_rsamtools.R
\name{run_bgzip_rsamtools}
\alias{run_bgzip_rsamtools}
\title{Run bgzip: Rsamtools}
\usage{
run_bgzip_rsamtools(target_path, bgz_file, force_new = TRUE, verbose = TRUE)
}
\arguments{
\item{target_path}{Path to full GWAS/QTL summary statistics file.}

\item{bgz_file}{Path to resulting bgz-compressed file after
\code{target_path} has been converted to tabix format.}

\item{force_new}{Force the creation of a 
new bgzip file (\emph{.bgz}) and a new tabix index file (\emph{.tbi}).}

\item{verbose}{Print messages.}
}
\description{
Support function for \link[echotabix]{run_bgzip}.
}
\keyword{internal}
