% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gunzip.R
\name{run_gunzip}
\alias{run_gunzip}
\title{Run gunzip}
\usage{
run_gunzip(
  path,
  conda_env = "echoR_mini",
  method = c("R.utils", "conda"),
  outputs = c("command", "path", "data"),
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to file.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{method}{Method to gunzip \code{path} with:
\describe{
\item{"conda": }{Finds gunzip binary in the specified 
\code{conda_env} and uses it.}
\item{"R.utils": }{Uses the \link[R.utils]{gunzip} function.}
}}

\item{outputs}{Which outputs to return in a list. 
Can be one or more of the following:
\describe{
\item{"command" : }{The command run to gunzip the file.}
\item{"path" : }{Path to the gunzipped file.}
\item{"data" : }{The data imported from the gunzipped file.}
}
If only one valid option is selected, the item will be returned without
being embedded within a list.}

\item{verbose}{Print messages.}
}
\description{
Decompress a file using \code{gunzip}.
}
\examples{
\dontrun{
dat <- echodata::BST1
tmp <- tempfile(fileext = ".csv.gz")
data.table::fwrite(dat, tmp)
out <- echotabix::run_gunzip(path=tmp)
}
}
