% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gunzip_conda.R
\name{run_gunzip_conda}
\alias{run_gunzip_conda}
\title{Run gunzip: conda}
\usage{
run_gunzip_conda(
  path,
  gunzip_ex = NULL,
  outputs = c("command", "path", "data"),
  save_path = gsub(".gz|.bgz", "", path),
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to file.}

\item{outputs}{\describe{
\item{"command"}{Text string of the command (without executing it.)}
\item{"path"}{Path to the saved data file.}
\item{"data"}{The resulting data in \link[data.table]{data.table} format.}
}}

\item{save_path}{File to save the results to.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{verbose}{Print messages.}
}
\description{
Support function for \link[echotabix]{run_gunzip}.
}
\keyword{internal}
