% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_vcf.R
\name{save_vcf}
\alias{save_vcf}
\title{Save VCF}
\usage{
save_vcf(vcf, query_save, save_path, verbose = TRUE)
}
\arguments{
\item{query_save}{Whether to save the results of the query on disk. 
\emph{Note}: Writing to disk can take some time.}

\item{save_path}{File path to save query subset to (as VCF).}

\item{verbose}{Print messages.}
}
\value{
save_path
}
\description{
Save a VCF file to disk.
}
\keyword{internal}
