% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_method.R
\name{select_method}
\alias{select_method}
\title{Select method}
\usage{
select_method(
  fn,
  fn_arg = "method",
  method,
  rhtslib_pkgs = c("variantannotation", "rsamtools", "rtracklayer"),
  verbose = TRUE
)
}
\arguments{
\item{fn}{Unevaluated function to select method for.}

\item{fn_arg}{Name of the argument that lists the various methods available
for that particular function.}

\item{method}{Method selected by user.}

\item{rhtslib_pkgs}{List of R packages that depend on \code{Rhtslib}.}

\item{verbose}{Print messages.}
}
\value{
A valid method.
}
\description{
Select a valid alternative method when the one chosen is not compatible with
the current Bioconductor version.
When provided a valid method, simply reutrns that method.
}
\keyword{internal}
