% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_coordinates.R
\name{sort_coordinates}
\alias{sort_coordinates}
\title{Sort coordinates}
\source{
\href{https://www.systutorials.com/docs/linux/man/1-tabix/}{
sorting instructions}

\href{https://unix.stackexchange.com/a/382801}{
StackExchange discussion of how \code{sort} handles numbers}
}
\usage{
sort_coordinates(
  target_path,
  chrom_col,
  start_col,
  end_col = start_col,
  comment_char = NULL,
  save_path = NULL,
  method = c("bash", "data.table"),
  outputs = c("command", "path", "data"),
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to full GWAS/QTL summary statistics file.}

\item{chrom_col}{Name of the chromosome column 
in the \code{target_path} file.}

\item{start_col}{Name of the genomic start position column 
in the \code{target_path} file.}

\item{end_col}{Name of the genomic end position column 
in the \code{target_path} file.}

\item{comment_char}{A single character which, when present as 
the first character in a line, indicates that the line is to be omitted 
from indexing.}

\item{save_path}{File to save the results to.}

\item{method}{Method to sort coordinates with.}

\item{outputs}{\describe{
\item{"command"}{Text string of the command (without executing it.)}
\item{"path"}{Path to the saved data file.}
\item{"data"}{The resulting data in \link[data.table]{data.table} format.}
}}

\item{verbose}{Print messages.}
}
\description{
Sort a table of summary statistics by their genomic coordinates 
without reading the file into R.
}
\examples{
\dontrun{
dat <- echodata::BST1
tmp <- tempfile()
data.table::fwrite(dat, tmp)
out <- echotabix::sort_coordinates(target_path=tmp,
                                   chrom_col = "CHR",
                                   start_col = "POS")
}
}
