% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_common_snps.R
\name{subset_common_snps}
\alias{subset_common_snps}
\title{Subset LD matrix and dataframe to only their shared SNPs}
\usage{
subset_common_snps(LD_matrix, dat, fillNA = 0, verbose = F)
}
\value{
data.frame
}
\description{
Find the SNPs that are shared between an LD matrix and
another data.frame with a `SNP` column.
Then remove any non-shared SNPs from both objects.
}
\details{
\code{
BST1 <- echodata::BST1
BST1_LD_matrix <- echodata::BST1_LD_matrix
dat <- BST1
out <- echoLD:::subset_common_snps(LD_matrix = BST1_LD_matrix, dat = BST1)
}
}
\concept{SNP filters}
\keyword{internal}
