% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffixes.R
\name{suffixes}
\alias{suffixes}
\title{List supported file formats}
\usage{
suffixes(
  tabular = TRUE,
  tabular_compressed = TRUE,
  vcf = TRUE,
  vcf_compressed = TRUE
)
}
\arguments{
\item{tabular}{Include tabular formats.}

\item{tabular_compressed}{Include compressed tabular formats.}

\item{vcf}{Include Variant Call Format.}

\item{vcf_compressed}{Include compressed Variant Call Format.}
}
\value{
File formats
}
\description{
List supported file formats
}
\keyword{internal}
