% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix_path.R
\name{tabix_path}
\alias{tabix_path}
\title{Construct tabix path}
\usage{
tabix_path(path, study_dir = NULL)
}
\arguments{
\item{path}{Path to file.}

\item{study_dir}{[optional] Path to study-specific subfolder.}
}
\description{
Given some summary stats file, 
construct a name for the resulting sorted, bgzip-compressed tabix file.
}
\examples{
bgz_file = echotabix::tabix_path(path = "mysumstatsfile.tsv.gz") 
}
\seealso{
Other tabix functions: 
\code{\link{convert}()},
\code{\link{index}},
\code{\link{query_table}()},
\code{\link{query_vcf}()},
\code{\link{read_bgz}()},
\code{\link{run_bgzip}()},
\code{\link{vcf_path}()}
}
\concept{tabix functions}
