% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_path.R
\name{vcf_path}
\alias{vcf_path}
\title{VCF path}
\usage{
vcf_path(
  target_path,
  query_granges = NULL,
  locus_dir = tempdir(),
  subdir = "VCF",
  use_coord_prefix = TRUE,
  whole_vcf = FALSE
)
}
\arguments{
\item{target_path}{Path to local VCF file or remote URL.}

\item{locus_dir}{Locus-specific folder.}

\item{use_coord_prefix}{Add min/max genomic coordinates 
(e.g. "chr4-14737349-16737284") to the file name.}
}
\description{
Construct the \code{save_path} to VCF subset 
extracted by \link[echotabix]{query_vcf}.
}
\examples{
path <- system.file("extdata", "BST1.1KGphase3.vcf.bgz",
                    package = "echotabix")
locus_dir <- echodata::locus_dir
BST1 <- echodata::BST1
save_path <- echotabix::vcf_path(query_dat = query_dat,
                                 locus_dir = locus_dir,
                                 path = path)
}
\seealso{
Other tabix functions: 
\code{\link{convert}()},
\code{\link{index}},
\code{\link{query_table}()},
\code{\link{query_vcf}()},
\code{\link{read_bgz}()},
\code{\link{run_bgzip}()},
\code{\link{tabix_path}()}
}
\concept{tabix functions}
