% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_delimiter.R
\name{check_delimiter}
\alias{check_delimiter}
\title{Check delimiter}
\usage{
check_delimiter(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to file.}

\item{verbose}{Print messages.}
}
\description{
Automatically infer what delimiter is used to separate columns 
(e.g. ",", "\\t"," "). Uses messages from \link[data.table]{fread} to 
extract this inference.
}
\examples{
dat <- echodata::BST1
path <- tempfile()
data.table::fwrite(dat, path, sep="\t")
delim <- echotabix::check_delimiter(path=path)
}
