% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tabix_path.R
\name{construct_tabix_path}
\alias{construct_tabix_path}
\title{Construct tabix path}
\usage{
construct_tabix_path(target_path, study_dir = NULL)
}
\arguments{
\item{target_path}{Path to full GWAS/QTL summary statistics file.}

\item{study_dir}{[optional] Path to study-specific subfolder.}
}
\description{
Given some summary stats file, 
construct a name for the resulting sorted, bgzip-compressed tabix file.
}
\examples{
bgz_file = echotabix::construct_tabix_path(
    target_path = "mysumstatsfile.tsv.gz")
}
\seealso{
Other tabix functions: 
\code{\link{construct_vcf_path}()},
\code{\link{convert}()},
\code{\link{index}},
\code{\link{query_table}()},
\code{\link{query_vcf}()},
\code{\link{read_bgz}()},
\code{\link{run_bgzip}()}
}
\concept{tabix functions}
