% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert summary stats file to tabix format}
\usage{
convert(fullSS_path, chrom_col = "CHR", position_col = "POS", verbose = TRUE)
}
\description{
Convert summary stats file to tabix format
}
\examples{
\dontrun{
fullSS_path <- echolocatoR::example_fullSS()
fullSS_tabix <- convert(fullSS_path = fullSS_path, position_col = "BP")
}
}
\seealso{
Other tabix: 
\code{\link{convert_and_query}()}
}
\concept{tabix}
