% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert to tabix}
\source{
\href{https://github.com/samtools/htslib/issues}{
samtools/tabix GitHub repo}

\href{https://github.com/Bioconductor/Rhtslib/issues/4}{
Rhtslib (which Rsamtools and seqminer depend on for tabix) 
is very out of date (1.7 vs. 1.15)}.

\href{https://github.com/bioconda/bioconda-recipes/issues/23404}{
\code{conda install tabix} was recently fixed so it installs htslib}
}
\usage{
convert(
  target_path,
  bgz_file = construct_tabix_path(target_path = target_path),
  chrom_col = "CHR",
  start_col = "POS",
  end_col = start_col,
  comment_char = NULL,
  format = NULL,
  convert_methods = list(sort_coordinates = "bash", run_bgzip = "Rsamtools", index =
    "Rsamtools"),
  conda_env = "echoR_mini",
  force_new = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to full GWAS/QTL summary statistics file.}

\item{bgz_file}{Path to resulting bgz-compressed file after
\code{target_path} has been converted to tabix format.}

\item{chrom_col}{Name of the chromosome column 
in the \code{target_path} file.}

\item{start_col}{Name of the genomic start position column 
in the \code{target_path} file.}

\item{end_col}{Name of the genomic end position column 
in the \code{target_path} file.}

\item{comment_char}{A single character which, when present as 
the first character in a line, indicates that the line is to be omitted 
from indexing.}

\item{format}{Format of the \code{target_path} file: "vcf" or "table".}

\item{convert_methods}{A named list containing methods to run each step with.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{force_new}{Force the creation of a 
new bgzip file (\emph{.bgz}) and a new tabix index file (\emph{.tbi}).}

\item{verbose}{Print messages.}
}
\description{
Convert a tabular file to compressed (bgzip), indexed, tabix format 
for rapid querying.
}
\examples{
#### Example with full data ####
# tmp <- echodata::example_fullSS()
#### Example with single locus ####
dat <- echodata::BST1
tmp <- tempfile()
data.table::fwrite(dat, tmp, sep="\t")
 
tabix_files <- echotabix::convert(target_path = tmp)
}
\seealso{
Other tabix functions: 
\code{\link{construct_tabix_path}()},
\code{\link{construct_vcf_path}()},
\code{\link{index}},
\code{\link{query_table}()},
\code{\link{query_vcf}()},
\code{\link{read_bgz}()},
\code{\link{run_bgzip}()}
}
\concept{tabix functions}
