% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_and_query.R
\name{convert_and_query}
\alias{convert_and_query}
\title{Convert and query}
\usage{
convert_and_query(
  fullSS_path,
  study_dir = NULL,
  subset_path = tempfile(".tsv.gz"),
  is_tabix = FALSE,
  chrom_col = "CHR",
  position_col = "BP",
  min_POS,
  max_POS,
  chrom,
  save_subset = TRUE,
  nThread = 1,
  conda_env = "echoR",
  verbose = TRUE
)
}
\value{
data.table of locus subset summary statistics
}
\description{
If it is not tabix format already
(determined by checking for a \link{.tbi}
file of the same name in the same directory),
the full summary statistics file is converted into tabix format
for super fast querying.
A query is then made using the min/max genomic positions to extract a
 locus-specific summary stats file.
}
\examples{
\dontrun{
data("BST1")
fullSS_path <- echolocatoR::example_fullSS()

subset_path <- file.path(tempdir(), "BST1_Nalls23andMe_2019_subset.tsv.gz")
dat <- convert_and_query(
    fullSS_path = fullSS_path,
    subset_path = subset_path,
    min_POS = min(BST1$POS),
    max_POS = max(BST1$POS),
    chrom = BST1$CHR[1]
)
}
}
\seealso{
Other tabix: 
\code{\link{convert}()}
}
\concept{tabix}
