% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_vcf.R
\name{dt_to_vcf}
\alias{dt_to_vcf}
\title{\link[data.table]{data.table} to VCF}
\usage{
dt_to_vcf(
  dat,
  save_path = tempfile(fileext = "_converted.vcf"),
  tabix_index = FALSE,
  nThread = 1,
  ...
)
}
\arguments{
\item{dat}{data.frame to convert to VCF file.}

\item{save_path}{File path to save formatted data. Defaults to
\code{tempfile(fileext=".tsv.gz")}.}

\item{tabix_index}{Index the formatted summary statistics with
\href{http://www.htslib.org/doc/tabix.html}{tabix} for fast querying.}

\item{nThread}{The number of threads to use. Experiment to see what works best for your data on your hardware.}

\item{...}{Additional arguments passed to 
\link[MungeSumstats]{standardise_header}.}
}
\description{
Convert a \link[data.table]{data.table} to a VCF file
Used to be performed with 
\href{https://github.com/RajLabMSSM/echolocatoR/blob/0ccf40d2f126f755074e731f82386e4e01d6f6bb/R/dataframe_2_vcf.R}{
\code{bcftools convert}, but \code{MungeSumstats} works much better}.
\strong{WARNING: } This method only works for a
 \link[data.table]{data.table} with a single sample. 
 It cannot parse multiple pieces of information stored in the same column.
}
\examples{
save_path <- echotabix::dt_to_vcf(dat=echodata::BST1)
}
