% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Tabix-index a file}
\usage{
index(
  bgz_file,
  chrom_col,
  start_col,
  end_col = start_col,
  comment_char = NULL,
  force_new = TRUE,
  method = c("conda", "seqminer", "Rsamtools"),
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{bgz_file}{Path to a file that has been compressed with \code{bgzip}
(e.g. via \link[echotabix]{run_bgzip}).}

\item{comment_char}{Comment character denoting which row contains the column 
names (e.g. "#CHR" or "SNP").}

\item{method}{Method to index tabix file with.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{verbose}{Print messages.}
}
\description{
Tabix-index a tabular summary statistics file.
}
\examples{
dat <- echodata::BST1
tmp <- tempfile(fileext = ".tsv.gz")
data.table::fwrite(dat, tmp, sep="\t")
bgz_file <- echotabix::run_bgzip(fullSS_path = tmp,
                                 chrom_col = "CHR",
                                 start_col = "POS")
tbi_file <- echotabix::index(bgz_file = bgz_file, 
                             chrom_col = "CHR", 
                             start_col = "POS")
}
\seealso{
Other tabix functions: 
\code{\link{convert}()},
\code{\link{query_table}()},
\code{\link{query_vcf}()},
\code{\link{read_bgz}()},
\code{\link{run_bgzip}()},
\code{\link{tabix_path}()},
\code{\link{vcf_path}()}
}
\concept{tabix functions}
