% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.r
\name{liftover}
\alias{liftover}
\title{Genome build liftover}
\source{
\href{https://doi.org/doi:10.18129/B9.bioc.liftOver}{liftOver}

\href{https://hgdownload.cse.ucsc.edu/goldenpath/hg19/liftOver/}{
UCSC chain files}
}
\usage{
liftover(
  sumstats_dt,
  convert_ref_genome = NULL,
  ref_genome = NULL,
  chrom_col = "CHR",
  start_col = "POS",
  end_col = start_col,
  as_granges = FALSE,
  style = "NCBI",
  verbose = TRUE
)
}
\arguments{
\item{sumstats_dt}{data table of the GWAS/QTL summary statistics.}

\item{convert_ref_genome}{name of the reference genome to convert to
("GRCh37" or "GRCh38"). This will only occur if the current genome build does
not match. Default is not to convert the genome build (\code{NULL}).}

\item{ref_genome}{name of the reference genome used for the GWAS ("GRCh37" or
"GRCh38"). Argument is case-insensitive.
 Default is \code{NULL} which infers the reference genome from the data.}

\item{chrom_col}{column for chromosome}

\item{start_col}{column for start position}

\item{end_col}{column for end position (is the same as start for snps)}

\item{as_granges}{Return results as \link[GenomicRanges]{GRanges}.}

\item{style}{\link[GenomicRanges]{GRanges} style
(e.g. "UCSC" = "chr4; "NCBI" = 4).}

\item{verbose}{Print messages.}
}
\value{
Lifted summary stats \code{data.table}
}
\description{
Transfer genomic coordinates from one genome build to another.
}
