% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{liftover}
\alias{liftover}
\title{Genome build liftover}
\source{
\href{https://doi.org/doi:10.18129/B9.bioc.liftOver}{liftOver}

\href{https://hgdownload.cse.ucsc.edu/goldenpath/hg19/liftOver/}{
UCSC chain files}
}
\usage{
liftover(
  dat,
  chrom_col = "CHR",
  start_col = "POS",
  end_col = start_col,
  build_conversion = c("hg38ToHg19", "hg19ToHg38"),
  as_granges = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{SNP-level data table.}

\item{chrom_col}{Name of the chromosome column.}

\item{start_col}{Name of the start position column.}

\item{end_col}{Name of the end position column
(can be same as \code{start_col} if all data is SNP-level).}

\item{build_conversion}{From which to which genome build
to lift over \code{dat}.}

\item{as_granges}{Return lifted \code{dat} as
\link[GenomicRanges]{GenomicRanges} object.}

\item{verbose}{Print messages.}
}
\description{
Transfer genomic coordinates from one genome build to another.
}
\examples{
data("BST1")
dat_lifted <- liftover(dat = BST1, build_conversion = "hg19ToHg38")
}
