% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query tabix}
\usage{
query(
  target_path,
  target_format = NULL,
  query_dat,
  query_granges = construct_query(query_dat = query_dat, query_chrom_col = "CHR",
    query_start_col = "POS", query_snp_col = "SNP"),
  samples = character(),
  query_save = TRUE,
  query_save_path = tempfile(fileext = ".gz"),
  target_genome = "GRCh37",
  query_genome = "GRCh37",
  query_method = c("rsamtools", "seqminer", "conda"),
  conda_env = "echoR",
  query_force_new = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to full GWAS/QTL summary statistics file.}

\item{query_dat}{Variant-level summary statistics.}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{query_save}{Whether to save the queried data subset.}

\item{target_genome}{Genome build of the VCF file.}

\item{nThread}{Number of threads to use.}

\item{verbose}{Print messages.}
}
\description{
Query a tabix table or VCF.
}
\examples{
query_dat <- echodata::BST1

#### local ####
target_path <- echodata::example_fullSS()
tabix_files <- echotabix::convert(target_path = target_path, 
                                  start_col = "BP")
query_res <- echotabix::query(
    target_path = tabix_files$path,
    query_dat = query_dat)
}
