% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query tabix}
\usage{
query(
  target_path,
  target_index = paste0(target_path, ".tbi"),
  target_format = NULL,
  query_granges,
  samples = character(),
  query_save = TRUE,
  query_save_path = tempfile(fileext = ".gz"),
  target_genome = "GRCh37",
  query_genome = "GRCh37",
  query_method = c("rsamtools", "variantannotation", "conda", "seqminer"),
  conda_env = "echoR_mini",
  query_force_new = FALSE,
  as_datatable = TRUE,
  overlapping_only = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to full GWAS/QTL summary statistics file.}

\item{target_index}{Tabix index file for \code{target_path}.}

\item{target_format}{Format of the \code{target_path} file: "vcf" or "table".}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{query_save}{Whether to save the queried data subset.}

\item{query_save_path}{Path to save retrieved query subset to.}

\item{target_genome}{Genome build of the VCF file.}

\item{query_genome}{Genome build that the \code{query_granges} is aligned to.}

\item{query_method}{Method used for querying. 
See \link[echotabix]{query} for available options.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{query_force_new}{If the query subset (\code{query_save_path})
already exists, set \code{query_force_new=TRUE}
 to retrieve a new query subset.}

\item{as_datatable}{Return the VCF subset 
file as a \link[data.table]{data.table} 
(using \link[echotabix]{vcf_to_dt}).
If \code{save_path=TRUE} the file will still be saved 
as a bgzip-compressed VCF file.}

\item{overlapping_only}{Remove variants that do not overlap with the 
positions in \code{query_dat}.}

\item{nThread}{Number of threads to use.}

\item{verbose}{Print messages.}
}
\description{
Query a tabix table or VCF.
}
\examples{
query_dat <- echodata::BST1

#### local ####
target_path <- echodata::example_fullSS()
tabix_files <- echotabix::convert(target_path = target_path, 
                                  start_col = "BP")
query_res <- echotabix::query(
    target_path = tabix_files$path,
    query_granges = query_dat)
}
