% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query tabix}
\usage{
query(
  target_path,
  target_format = NULL,
  query_dat = NULL,
  query_chrom_col = "CHR",
  query_start_col = "POS",
  query_end_col = query_start_col,
  query_snp_col = "SNP",
  query_chrom = NULL,
  query_start_pos = NULL,
  query_end_pos = query_start_pos,
  samples = character(),
  query_save = TRUE,
  locus_dir = tempdir(),
  query_save_path = tempfile(fileext = ".gz"),
  target_genome = "GRCh37",
  query_genome = "GRCh37",
  query_method = c("rsamtools", "seqminer", "conda"),
  conda_env = "echoR",
  query_force_new = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to local VCF file or remote URL.}

\item{query_dat}{Variant-level summary statistics.}

\item{query_chrom_col}{Name of the chromosome column in \code{query_dat} (e.g. "CHR").}

\item{query_start_col}{Name of the starting genomic position
column in \code{query_dat} (e.g. "POS","start").}

\item{query_end_col}{Name of the ending genomic position
column in \code{query_dat} (e.g. "POS","end"). 
Can be the same as \code{query_start_col} when \code{query_dat} only contains SNPs that
span 1 base pair (bp) each.}

\item{query_chrom}{Which chromosome to query (e.g. 1 or "chr1").}

\item{query_start_pos}{The first position to query.}

\item{query_end_pos}{The last position to query.}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{query_save}{Whether to save the queried data subset.}

\item{locus_dir}{Locus-specific folder.}

\item{target_genome}{Genome build of the VCF file.}

\item{nThread}{Number of threads to use.}

\item{verbose}{Print messages.}
}
\description{
Query a tabix table or VCF.
}
\examples{
dat <- echodata::BST1

#### local ####
fullSS_path <- echodata::example_fullSS()
tabix_files <- echotabix::convert(fullSS_path = fullSS_path, 
                                  start_col = "BP")
query_res <- echotabix::query(
    target_path = tabix_files$data,
    query_chrom = dat$CHR[1],
    query_start_pos = min(dat$POS),
    query_end_pos = max(dat$POS))
}
