% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table.R
\name{query_table}
\alias{query_table}
\title{Query a tabix file}
\usage{
query_table(
  target_path,
  query_chrom,
  query_start_pos,
  query_end_pos = query_start_pos,
  method = c("rsamtools", "seqminer", "conda"),
  local = NULL,
  query_save = TRUE,
  save_path = tempfile(fileext = "tsv.gz"),
  conda_env = "echoR",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{query_chrom}{Which chromosome to query (e.g. 1 or "chr1").}

\item{query_start_pos}{The first position to query.}

\item{query_end_pos}{The last position to query.}

\item{local}{Whether \code{target_path} is stored locally or 
on a remote server/website.
By default (\code{NULL}) will infer local status and 
use the appropriate method.}

\item{verbose}{Print messages.}
}
\value{
\code{data.table} with the queried subset of genomic data.
}
\description{
Query by genomic coordinates.
}
\examples{
dat <- echodata::BST1

#### local ####
fullSS_path <- echodata::example_fullSS()
tabix_files <- echotabix::convert(fullSS_path = fullSS_path, 
                                  start_col = "BP")
query_res <- echotabix::query_table(
    target_path = tabix_files$data,
    query_chrom = dat$CHR[1],
    query_start_pos = min(dat$POS),
    query_end_pos = max(dat$POS)
)

#### remote ####
target_path <- file.path(
    "https://egg2.wustl.edu/roadmap/data/byFileType",
    "chromhmmSegmentations/ChmmModels/coreMarks/jointModel/final",
    "E099_15_coreMarks_dense.bed.bgz"
)
tab <- echotabix::query_table(
    target_path = target_path,
    query_chrom = dat$CHR[1],
    query_start_pos = min(dat$POS),
    query_end_pos = max(dat$POS)
) 
}
\seealso{
Other tabix functions: 
\code{\link{convert}()},
\code{\link{index}},
\code{\link{query_vcf}()},
\code{\link{read_bgz}()},
\code{\link{run_bgzip}()},
\code{\link{tabix_path}()},
\code{\link{vcf_path}()}
}
\concept{tabix functions}
