% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table.R
\name{query_table}
\alias{query_table}
\title{Query a tabix file}
\usage{
query_table(
  target_path,
  query_dat,
  query_granges = construct_query(query_dat = query_dat, query_chrom_col = "CHR",
    query_start_col = "POS", query_snp_col = "SNP"),
  method = c("rsamtools", "seqminer", "conda"),
  local = NULL,
  query_save = TRUE,
  save_path = tempfile(fileext = "tsv.gz"),
  conda_env = "echoR",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{query_dat}{Variant-level summary statistics.}

\item{local}{Whether \code{target_path} is stored locally or 
on a remote server/website.
By default (\code{NULL}) will infer local status and 
use the appropriate method.}

\item{verbose}{Print messages.}
}
\value{
\code{data.table} with the queried subset of genomic data.
}
\description{
Query by genomic coordinates.
}
\examples{
query_dat <- echodata::BST1

#### local ####
fullSS_path <- echodata::example_fullSS()
tabix_files <- echotabix::convert(fullSS_path = fullSS_path, 
                                  start_col = "BP")
query_res <- echotabix::query_table(
    target_path = tabix_files$data,
    query_dat = query_dat)

#### remote ####
target_path <- file.path(
    "https://egg2.wustl.edu/roadmap/data/byFileType",
    "chromhmmSegmentations/ChmmModels/coreMarks/jointModel/final",
    "E099_15_coreMarks_dense.bed.bgz"
)
query_res2 <- echotabix::query_table(
    target_path = target_path,
    query_dat = query_dat) 
}
\seealso{
Other tabix functions: 
\code{\link{convert}()},
\code{\link{index}},
\code{\link{query_vcf}()},
\code{\link{read_bgz}()},
\code{\link{run_bgzip}()},
\code{\link{tabix_path}()},
\code{\link{vcf_path}()}
}
\concept{tabix functions}
