% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table_conda.R
\name{query_table_conda}
\alias{query_table_conda}
\title{Query tabular: conda}
\usage{
query_table_conda(
  target_path,
  query_granges,
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{verbose}{Print messages.}
}
\description{
Uses a conda-based installation of tabix instead of compiled 
\code{C} code from \pkg{Rhtslib}.
}
\keyword{internal}
