% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table_conda.R
\name{query_table_conda}
\alias{query_table_conda}
\title{Query tabular: conda}
\usage{
query_table_conda(
  target_path,
  query_granges,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{verbose}{Print messages.}
}
\description{
Uses a conda-based installation of tabix instead of compiled 
\code{C} code from \pkg{Rhtslib}.
}
\keyword{internal}
