% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table_conda.R
\name{query_table_conda}
\alias{query_table_conda}
\title{Query tabular: conda}
\usage{
query_table_conda(
  target_path,
  query_chrom,
  query_start_pos,
  query_end_pos,
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{query_chrom}{Which chromosome to query (e.g. 1 or "chr1").}

\item{query_start_pos}{The first position to query.}

\item{query_end_pos}{The last position to query.}

\item{verbose}{Print messages.}
}
\description{
Uses a conda-based installation of tabix instead of compiled 
\code{C} code from \pkg{Rhtslib}.
}
\keyword{internal}
