% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table_rsamtools.R
\name{query_table_rsamtools}
\alias{query_table_rsamtools}
\title{Query tabular: \pkg{Rsamtools}}
\source{
\href{https://github.com/Bioconductor/Rsamtools/issues/8}{
 Bugs in \code{Rsamtools::scanTabix}}
}
\usage{
query_table_rsamtools(
  target_path,
  query_chrom,
  query_start_pos,
  query_end_pos = query_start_pos,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{query_chrom}{Which chromosome to query (e.g. 1 or "chr1").}

\item{query_start_pos}{The first position to query.}

\item{query_end_pos}{The last position to query.}

\item{verbose}{Print messages.}
}
\description{
\pkg{Rsamtools} can query both local and remote files. 
 However, it seems unable to read the header of 
 tabix files created with \pkg{seqminer}.
}
\keyword{internal}
