% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table_rsamtools.R
\name{query_table_rsamtools}
\alias{query_table_rsamtools}
\title{Query tabular: \pkg{Rsamtools}}
\source{
\href{https://github.com/Bioconductor/Rsamtools/issues/8}{
 Bugs in \code{Rsamtools::scanTabix}}
}
\usage{
query_table_rsamtools(target_path, query_granges, verbose = TRUE)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{verbose}{Print messages.}
}
\description{
\pkg{Rsamtools} can query both local and remote files. 
 However, it seems unable to read the header of 
 tabix files created with \pkg{seqminer}.
}
\keyword{internal}
