% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table_seqminer.R
\name{query_table_seqminer}
\alias{query_table_seqminer}
\title{Query tabular: \pkg{seqminer}}
\source{
\href{https://github.com/zhanxw/seqminer/issues/25}{
GitHub Issues: coordinate order error}

\href{https://github.com/zhanxw/seqminer/issues/20}{
GitHub Issues: remote file error}

\href{https://www.utsouthwestern.edu/labs/zhan/contact/}{
Lab contact details for \code{seqminer} maintainer}

\href{https://pubmed.ncbi.nlm.nih.gov/26394715/}{
\code{seqminer} publication}

\href{https://pubmed.ncbi.nlm.nih.gov/32756942/}{
\code{seqminer2} publication (same package and GitHub repo, just updated)}
}
\usage{
query_table_seqminer(target_path, query_granges, verbose = TRUE)
}
\arguments{
\item{target_path}{Path to tabix file.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{verbose}{Print messages.}
}
\description{
\strong{Limitations of seqminer}:
\itemize{
\item{Doesn't work with remote files.}
\item{Assumes header column always starts with "#" (which is often incorrect)
without the option for the user to specify otherwise. 
The only way that summary statistics can be used with \pkg{seqminer}
 is if they adhere to this convention, or were indexed using \pkg{seqminer}'s
 \link[seqminer]{tabix.createIndex} function.}
\item{Maintainers are unresponsive to requests for bug fixes.}
}  
\strong{Advantages of seqminer}:
\itemize{
\item{\link[seqminer]{tabix.read.table} automatically converts query results
to data.frame format (though this can now also be done by 
\link[echotabix]{scanTabix_to_dt}).}
} 
\pkg{seqminer} appears to be maintained to some degree
 (based on the latest commits), but the maintainers
have been unresponsive to bug reports for years. 
This limits the consistent usability of \pkg{seqminer}.
}
\keyword{internal}
