% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tabular_rsamtools.R
\name{query_tabular_rsamtools}
\alias{query_tabular_rsamtools}
\title{query_tabular_rsamtools}
\usage{
query_tabular_rsamtools(
  fullSS_tabix,
  chrom,
  start_pos,
  end_pos,
  verbose = TRUE
)
}
\description{
query_tabular_rsamtools
}
\keyword{internal}
