% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{query_vcf}
\alias{query_vcf}
\title{Query VCF file}
\usage{
query_vcf(
  dat,
  vcf_url,
  locus_dir = file.path(tempdir(), "LD"),
  vcf_name = gsub(".vcf|.gz.|bgz", "", basename(vcf_url)),
  ref_genome = "GRCh37",
  samples = NULL,
  force_new_vcf = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{SNP-level data table.}

\item{vcf_url}{URL or path to VCF.}

\item{locus_dir}{Directory to store LD in.}

\item{vcf_name}{VCF reference name (e.g. "1KGphase1").}

\item{ref_genome}{Genome build of the VCF file.}

\item{samples}{Sample names to subset the VCF by before computing LD.}

\item{force_new_vcf}{Force the creation of a new LD file even if one exists.}

\item{verbose}{Print messages.}
}
\value{
\link[VariantAnnotation]{VCF} object.
}
\description{
Query a (remote) Variant Call Format (VCF) file.
}
\seealso{
Other LD: 
\code{\link{construct_subset_vcf_name}()},
\code{\link{get_locus_vcf_folder}()}
}
\concept{LD}
