% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{query_vcf}
\alias{query_vcf}
\title{Query VCF file}
\usage{
query_vcf(
  target_path,
  target_genome = "GRCh37",
  query_granges,
  samples = character(),
  method = c("variantannotation", "conda", "rtracklayer", "seqminer"),
  overlapping_only = FALSE,
  query_save = TRUE,
  save_path = construct_vcf_path(target_path = target_path, query_granges =
    query_granges),
  force_new = FALSE,
  as_datatable = FALSE,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to local VCF file or remote URL.}

\item{target_genome}{Genome build of the VCF file.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{method}{Method to query VCF with.}

\item{overlapping_only}{Remove variants that do not overlap with the 
positions in \code{query_dat}.}

\item{query_save}{Whether to save the results of the query on disk. 
\emph{Note}: Writing to disk can take some time.}

\item{save_path}{File path to save query subset to (as VCF).}

\item{force_new}{Force the creation of a new VCF subset file
even if one exists.}

\item{as_datatable}{Return the VCF subset 
file as a \link[data.table]{data.table} 
(using \link[echotabix]{vcf_to_dt}).
If \code{save_path=TRUE} the file will still be saved 
as a bgzip-compressed VCF file.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{verbose}{Print messages.}
}
\value{
\link[VariantAnnotation]{VCF} object, 
or \link[data.table]{data.table} (when \code{as_datatable=TRUE}).
}
\description{
Query a Variant Call Format (VCF) file. 
The VCF file can be either local or remote.
}
\examples{
query_dat <- echodata::BST1
target_path <- system.file("extdata", "BST1.1KGphase3.vcf.bgz",
                    package = "echodata")

#### Import ####
vcf <-  query_vcf(
    query_granges = query_dat,
    target_path = target_path)
}
\seealso{
Other tabix functions: 
\code{\link{construct_tabix_path}()},
\code{\link{construct_vcf_path}()},
\code{\link{convert}()},
\code{\link{index}},
\code{\link{query_table}()},
\code{\link{read_bgz}()},
\code{\link{run_bgzip}()}
}
\concept{tabix functions}
