% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{query_vcf}
\alias{query_vcf}
\title{Query VCF file}
\usage{
query_vcf(
  target_path,
  target_genome = "GRCh37",
  query_dat,
  query_granges = construct_query(query_dat = query_dat, query_chrom_col = "CHR",
    query_start_col = "POS", query_snp_col = "SNP"),
  samples = character(),
  overlapping_only = FALSE,
  query_save = TRUE,
  save_path = construct_vcf_path(target_path = target_path, query_granges =
    query_granges),
  force_new = FALSE,
  as_datatable = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to local VCF file or remote URL.}

\item{target_genome}{Genome build of the VCF file.}

\item{query_dat}{Variant-level summary statistics.}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{overlapping_only}{Remove variants that do not overlap with the 
positions in \code{query_dat}.}

\item{query_save}{Whether to save the results of the query on disk. 
\emph{Note}: Writing to disk can take some time.}

\item{save_path}{Path to save VCF results in.}

\item{force_new}{Force the creation of a new VCF subset file
even if one exists.}

\item{as_datatable}{Return the VCF subset 
file as a \link[data.table]{data.table} 
(using \link[echotabix]{vcf_to_dt}).
If \code{save_path=TRUE} the file will still be saved 
as a bgzip-compressed VCF file.}

\item{verbose}{Print messages.}
}
\value{
\link[VariantAnnotation]{VCF} object, 
or \link[data.table]{data.table} (when \code{as_datatable=TRUE}).
}
\description{
Query a Variant Call Format (VCF) file. 
The VCF file can be either local or remote.
}
\examples{
query_dat <- echodata::BST1
target_path <- system.file("extdata", "BST1.1KGphase3.vcf.bgz",
                    package = "echotabix"
)

#### Import ####
vcf <- echotabix::query_vcf(
    query_dat = query_dat,
    target_path = target_path, 
)
}
\seealso{
Other tabix functions: 
\code{\link{construct_tabix_path}()},
\code{\link{construct_vcf_path}()},
\code{\link{convert}()},
\code{\link{index}},
\code{\link{query_table}()},
\code{\link{read_bgz}()},
\code{\link{run_bgzip}()}
}
\concept{tabix functions}
