% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf_seqminer.R
\name{query_vcf_seqminer}
\alias{query_vcf_seqminer}
\title{Query VCF: \pkg{seqminer}}
\source{
\href{https://github.com/zhanxw/seqminer/issues/26}{
seqminer::readVCFToListByRange and seqminer::readVCFToMatrixByRange
 do not work and instead cause Rstudio to crash.
}

\code{ 
query_dat <- echodata::BST1[seq(1, 50), ]
target_path <- paste(
    "ftp://ftp-trace.ncbi.nih.gov/1000genomes/ftp/release/20110521/",
    "ALL.chr4.phase1_release_v3.20101123.snps_indels_svs.genotypes.vcf.gz",
    sep="/"
)
vcf <- echotabix:::query_vcf_rtracklayer(
    target_path = target_path,
    query_granges = query_dat)
}
}
\usage{
query_vcf_seqminer(
  target_path,
  target_genome = "GRCh37",
  query_granges,
  samples = character(),
  query_save = FALSE,
  save_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to local VCF file or remote URL.}

\item{target_genome}{Genome build of the VCF file.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{query_save}{Whether to save the results of the query on disk. 
\emph{Note}: Writing to disk can take some time.}

\item{save_path}{File path to save query subset to (as VCF).}

\item{verbose}{Print messages.}
}
\value{
A variant x sample data.frame
}
\description{
Query a subset of a VCF file (remote or local) 
using \link[seqminer]{tabix.read.table}.
\strong{Advantages of \pkg{seqminer}:}
\itemize{
\item{Does not rely on \pkg{Rsamtools} or \pkg{Rhtslib}, 
which are very outdated and prone to breaking.}
}
\strong{Disadvantages of \pkg{rtracklayer}:}
\itemize{
\item{Unable to query a subset of samples, 
unlike \link[VariantAnnotation]{scanVcf}.}
\item{Unable to return results as a structured 
\link[VariantAnnotation]{CollapsedVCF} object.}
}
}
\keyword{internal}
