% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf_variantannotation.R
\name{query_vcf_variantannotation}
\alias{query_vcf_variantannotation}
\title{Query VCF}
\source{
\href{https://bioconductor.org/packages/devel/bioc/vignettes/TVTB/inst/doc/VcfFilterRules.html}{
VariantAnnotation filtering}
}
\usage{
query_vcf_variantannotation(
  vcf_url,
  dat,
  samples = character(),
  genome = "GRCh37",
  save_path = tempfile(fileext = ".vcf.bgz"),
  verbose = TRUE
)
}
\arguments{
\item{vcf_url}{URL or path to VCF.}

\item{dat}{SNP-level data table.}

\item{samples}{Sample names to subset the VCF by before computing LD.}

\item{genome}{A \code{character} or \code{Seqinfo} object.
    \itemize{
      \item{\code{character}:}{ Genome identifier as a single string or named 
        character vector. Names of the character vector correspond to 
        chromosome names in the file. This identifier replaces the genome 
        information in the VCF \code{Seqinfo} (i.e., \code{seqinfo(vcf)}).
        When not provided, \code{genome} is taken from the VCF file header.
      }
      \item{\code{Seqinfo}:}{ When \code{genome} is provided as a \code{Seqinfo}
        it is propagated to the VCF output. If seqinfo information can be 
        obtained from the file, 
        (i.e., seqinfo(scanVcfHeader(fl)) is not empty), the output
        \code{Seqinfo} is a product of merging the two.

        If a param (i.e., ScanVcfParam) is used in the call to \code{readVcf}, 
        the seqlevels  of the param ranges must be present in \code{genome}.
      }
    }
  }

\item{verbose}{Print messages.}
}
\description{
# Alternative methods
\code{
 #### Rsamtools ####
 tab <- Rsamtools::scanTabix(file = vcf_url, gr)
 #### Rtracklayer ####
 tab <- Rsamtools::TabixFile(vcf_url)
 open(tab)
 tab <- rtracklayer::import(tab, which=gr)
 #### seqminer ####
 tab <- seqminer::tabix.read(tabixFile = vcf_url,
                              tabixRange = coord_range)
 header <- VariantAnnotation::scanVcfHeader(file = path)
}
}
\examples{
\dontrun{
BST1 <- echodata::BST1
dat <- BST1[seq(1, 50), ]
vcf_url <- file.path(
    "ftp://ftp-trace.ncbi.nih.gov/1000genomes/ftp/release/20110521/",
    "ALL.chr4.phase1_release_v3.20101123.snps_indels_svs.genotypes.vcf.gz"
)
vcf <- query_vcf_variantannotation(
    vcf_url = vcf_url,
    dat = dat
)
}

}
\keyword{internal}
