% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf_variantannotation.R
\name{query_vcf_variantannotation}
\alias{query_vcf_variantannotation}
\title{Query VCF: \pkg{VariantAnnotation}}
\source{
\href{https://bioconductor.org/packages/devel/bioc/vignettes/TVTB/inst/doc/VcfFilterRules.html}{
\pkg{VariantAnnotation} filtering vignette}

\href{https://github.com/MRCIEU/gwasvcf/blob/master/R/query.r}{
\code{gwasvcf} GitHub repo
}

\code{
BST1 <- echodata::BST1
query_dat <- BST1[seq(1, 50), ]
target_path <- paste(
    "ftp://ftp-trace.ncbi.nih.gov/1000genomes/ftp/release/20110521/",
    "ALL.chr4.phase1_release_v3.20101123.snps_indels_svs.genotypes.vcf.gz",
    sep="/"
)
vcf <- echotabix:::query_vcf_variantannotation(
    target_path = target_path,
    query_granges = query_dat)
}
}
\usage{
query_vcf_variantannotation(
  target_path,
  target_genome = NULL,
  query_granges,
  samples = character(),
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to local VCF file or remote URL.}

\item{target_genome}{Genome build of the VCF file.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{verbose}{Print messages.}
}
\value{
\link[VariantAnnotation]{CollapsedVCF} object.
}
\description{
Query a subset of a VCF file (remote or local) 
using \link[VariantAnnotation]{readVcf}.
\strong{Advantages of \pkg{VariantAnnotation}:}
\itemize{
\item{Is at least as fast as \link[Rsamtools]{scanTabix}.}
\item{Can query a specific subset of samples, 
unlike \link[Rsamtools]{scanTabix} which queries all samples at once.}
\item{
Automatically imports query results as a
 \link[VariantAnnotation]{CollapsedVCF} object, which contain lots of 
organized information about the query data and can be further processed
using other functions from 
\pkg{VariantAnnotation} and \pkg{snpStats}. 
By contrast, \link[Rsamtools]{scanTabix} returns a raw list of strings
that must be parsed by the user.
}
}
}
\keyword{internal}
