% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bgzip.R
\name{run_bgzip}
\alias{run_bgzip}
\title{Run bgzip}
\usage{
run_bgzip(
  fullSS_path,
  chrom_col,
  start_col,
  end_col = start_col,
  comment_char = NULL,
  bgz_file = tabix_path(path = fullSS_path),
  sort_rows = TRUE,
  force_new = TRUE,
  method = c("Rsamtools", "conda"),
  conda_env = "echoR",
  validate = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fullSS_path}{Path to full GWAS/QTL summary statistics file.}

\item{sort_rows}{Sort rows by genomic coordinates.}

\item{force_new}{Force the creation of a 
new bgzip file (\emph{.bgz}) and a new tabix index file (\emph{.tbi}).}

\item{method}{A named list containing methods to run each step with.}

\item{validate}{Check that the bgzip file exists and can be read in as a 
\link[data.table]{data.table}.}

\item{verbose}{Print messages.}
}
\description{
Compress a file using bgzip.
}
\examples{
#### Example with full data ####
# tmp <- echodata::example_fullSS()
#### Example with single locus ####
dat <- echodata::BST1
tmp <- tempfile()
data.table::fwrite(dat, tmp)

bgz_file <-  echotabix::run_bgzip(fullSS_path=tmp, 
                                 chrom_col="CHR", 
                                 start_col="BP")
}
\seealso{
Other tabix functions: 
\code{\link{convert}()},
\code{\link{index}},
\code{\link{query_table}()},
\code{\link{query_vcf}()},
\code{\link{read_bgz}()},
\code{\link{tabix_path}()},
\code{\link{vcf_path}()}
}
\concept{tabix functions}
