% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bgzip_conda.R
\name{run_bgzip_conda}
\alias{run_bgzip_conda}
\title{Run bgzip: conda}
\usage{
run_bgzip_conda(
  fullSS_path,
  bgz_file = tabix_path(path = fullSS_path),
  chrom_col,
  start_col,
  end_col,
  comment_char = NULL,
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{fullSS_path}{Path to full GWAS/QTL summary statistics file.}

\item{verbose}{Print messages.}
}
\description{
Support function for \link[echotabix]{run_bgzip}.
}
\keyword{internal}
