% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bgzip_conda.R
\name{run_bgzip_conda}
\alias{run_bgzip_conda}
\title{Run bgzip: conda}
\usage{
run_bgzip_conda(
  target_path,
  bgz_file = construct_tabix_path(target_path = target_path),
  chrom_col,
  start_col,
  end_col,
  comment_char = NULL,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to full GWAS/QTL summary statistics file.}

\item{verbose}{Print messages.}
}
\description{
Support function for \link[echotabix]{run_bgzip}.
}
\keyword{internal}
