% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gunzip_rutils.R
\name{run_gunzip_rutils}
\alias{run_gunzip_rutils}
\title{Run gunzip: R.utils}
\usage{
run_gunzip_rutils(
  path,
  outputs = c("command", "path", "data"),
  overwrite = TRUE,
  remove = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to file.}

\item{verbose}{Print messages.}
}
\description{
Support function for \link[echotabix]{run_gunzip}.
}
\keyword{internal}
