% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanTabix_to_dt.R
\name{scanTabix_to_dt}
\alias{scanTabix_to_dt}
\title{scanTabix to data.table}
\usage{
scanTabix_to_dt(
  header,
  queries,
  add_query_names = TRUE,
  remove_duplicates = TRUE,
  sep = "\\t",
  verbose = TRUE
)
}
\arguments{
\item{header}{Header, from the output of \link[Rsamtools]{headerTabix}.}

\item{queries}{A named list of query results, 
from the output of \link[Rsamtools]{scanTabix}.}

\item{add_query_names}{Add the names of each query to a new column 
named 'query'.}

\item{remove_duplicates}{Remove any duplicated rows.
Set \code{add_query_names=FALSE} to prevent each unique range (e.g. SNP) 
from appearing in more than one row.}

\item{sep}{ The separator between columns. Defaults to the character in the set \code{[,\\t |;:]} that separates the sample of rows into the most number of lines with the same number of fields. Use \code{NULL} or \code{""} to specify no separator; i.e. each line a single character column like \code{base::readLines} does.}

\item{verbose}{Print messages.}
}
\description{
Convert the output of \link[Rsamtools]{scanTabix}
 into a \link[data.table]{data.table}. 
 Can handle tabix files that are missing column names, 
 and query results that only have a single row.
}
\examples{
fl <- system.file("extdata", "example.gtf.gz", package="Rsamtools",
                  mustWork=TRUE)
tbx <- Rsamtools::TabixFile(fl)

param <- GenomicRanges::GRanges(
    c("chr1", "chr2"),
    IRanges::IRanges(c(1, 1), width=100000))
queries <- Rsamtools::scanTabix(tbx, param=param)
header <- Rsamtools::headerTabix(fl)

#### Convert ####
query_dt <-  echotabix::scanTabix_to_dt(header = header,
                                        queries = queries) 
}
