% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_coordinates_bash.R
\name{sort_coordinates_bash}
\alias{sort_coordinates_bash}
\title{Sort coordinates: bash}
\usage{
sort_coordinates_bash(
  target_path,
  chrom_col,
  start_col,
  end_col = start_col,
  comment_char = NULL,
  save_path = NULL,
  outputs = c("command", "path", "data"),
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to full GWAS/QTL summary statistics file.}

\item{chrom_col}{Name of the chromosome column 
in the \code{target_path} file.}

\item{start_col}{Name of the genomic start position column 
in the \code{target_path} file.}

\item{end_col}{Name of the genomic end position column 
in the \code{target_path} file.}

\item{comment_char}{A single character which, when present as 
the first character in a line, indicates that the line is to be omitted 
from indexing.}

\item{save_path}{File to save the results to.}

\item{outputs}{\itemize{
\item{"command"}{Text string of the command (without executing it.)}
\item{"path"}{Path to the saved data file.}
\item{"data"}{The resulting data in \link[data.table]{data.table} format.}
}}

\item{verbose}{Print messages.}
}
\description{
Support function for \link[echotabix]{sort_coordinates}.
}
\keyword{internal}
