% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_coordinates_bash.R
\name{sort_coordinates_bash}
\alias{sort_coordinates_bash}
\title{Sort coordinates: bash}
\usage{
sort_coordinates_bash(
  target_path,
  chrom_col,
  start_col,
  end_col = start_col,
  comment_char = NULL,
  save_path = NULL,
  outputs = c("command", "path", "data"),
  verbose = TRUE
)
}
\arguments{
\item{target_path}{Path to full GWAS/QTL summary statistics file.}

\item{save_path}{File to save the results to.}

\item{verbose}{Print messages.}
}
\description{
Support function for \link[echotabix]{sort_coordinates}.
}
\keyword{internal}
