% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_dt.R
\name{vcf_to_dt}
\alias{vcf_to_dt}
\title{VCF to \link[data.table]{data.table}}
\usage{
vcf_to_dt(vcf, expand = TRUE, verbose = TRUE)
}
\arguments{
\item{vcf}{Variant Call Format (VCF) file imported into R 
as a \pkg{VariantAnnotation} 
\link[VariantAnnotation]{CollapsedVCF} or 
\link[VariantAnnotation]{ExpandedVCF} object.}

\item{expand}{Expand data into multiple columns using
\code{VariantAnnotation::expand}.}

\item{verbose}{Print messages.}
}
\value{
A \link[data.table]{data.table}.
}
\description{
Converts a \pkg{VariantAnnotation} \link[VariantAnnotation]{CollapsedVCF}
or \link[VariantAnnotation]{ExpandedVCF} object 
to a \link[data.table]{data.table}. 
 
Adapted from \code{MungeSumstats:::vcf2df}, which was adapted from
a Gist by 
\href{https://gist.github.com/zhujack/849b75f5a8305edaeca1001dfb9c3fe9}{
zhujack}.
}
\examples{
vcf_file <- system.file("extdata", "BST1.1KGphase3.vcf.bgz",
                        package = "echotabix") 
vcf <- VariantAnnotation::readVcf(file = vcf_file)
vcf_dt <- echotabix::vcf_to_dt(vcf = vcf)
}
